add search component logic to homepage container, comment out search component, add utils
