+++ src/main.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './styles/index.css';
 import './styles/App.css';
-import Search from './app/components/search.component.tsx';
+import Search from './containers/homepage-container.tsx';
 import NavBar from './app/components/navbar/navbar.component.tsx';
 import { ThemeSwitcher } from './app/components/theme-switcher.component.tsx';
